* Refactored code: Changed folder structure, normalized file names, removed unused files * Added chat with nicknames and "is typing" functionality
