+++ about.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <title>Trazos</title>
     <link rel="stylesheet" href="css/result.css">
-    <link rel="stylesheet" href="css/singlePageInfo.css">
+    <link rel="stylesheet" href="css/about.css">
     <link rel="stylesheet" href="css/jquery.fullPage.css">
     <link href='https://fonts.googleapis.com/css?family=Roboto+Mono:700,300' rel='stylesheet' type='text/css'>
 </head>
@@ -164,6 +164,6 @@
 
     <script src="js/jquery.min.js"></script>
     <script src="js/jquery.fullPage.min.js" charset="utf-8"></script>
-    <script src="js/singlePageInfo.js" charset="utf-8"></script>
+    <script src="js/about.js" charset="utf-8"></script>
 </body>
 </html>
