+++ index.html
@@ -19,10 +19,9 @@
         </div>
     </div>
 
-    <script src="../js/jquery.min.js"></script>
-    <script src="../js/p5.min.js"></script>
-    <script src="../js/gifshot.min.js"></script>
-    <script src="../js/index.js"></script>
-    <script src="../js/sketchIndex.js"></script>
+    <script src="../js/lib/jquery.min.js"></script>
+    <script src="../js/lib/p5.min.js"></script>
+    <script src="../js/lib/gifshot.min.js"></script>
+    <script src="../js/trazos/landingSketch.js"></script>
 </body>
 </html>
