+++ partials/sidebar.inc.html
@@ -1,0 +1,92 @@
+<!-- Sidebar -->
+<div id="sidebar">
+    <a id="btnHome" href="/">
+        <img src="../img/home.svg" />
+    </a>
+    <div class="sidebar_button_color" id="color">
+        <div class="drop_color">
+            <div class="colors-wrap">
+                <div class="color_button" data-color="#000000"></div>
+                <div class="color_button" data-color="#ffffff"></div>
+                <div class="color_button" data-color="#ff0000"></div>
+                <div class="color_button" data-color="#ff4800"></div>
+                <div class="color_button" data-color="#fff200"></div>
+                <div class="color_button" data-color="#00ff6d"></div>
+                <div class="color_button" data-color="#0049ff"></div>
+                <div class="color_button" data-color="#AE00ff"></div>
+            </div>
+        </div>
+    </div>
+    <div class="sidebar_weight_button" id="weight">
+        <div class="weight_point weight_point_tam02"></div>
+        <div class="drop_weight">
+            <div class="weight_wrap">
+                <div class="weight_button weight_button_01"></div>
+                <div class="weight_button weight_button_02"></div>
+                <div class="weight_button weight_button_03"></div>
+                <div class="weight_button weight_button_04"></div>
+            </div>
+        </div>
+    </div>
+    <div class="sidebar_button">
+        <div class="wrapper wrapperSimple" id="loop">
+            <img src="../img/girar_1.svg" alt="">
+        </div>
+    </div>
+    <div class="sidebar_button">
+        <div class="wrapper wrapperDual wrapperHorizontal" id="velocidad" data-state="5" data-max-state="10">
+            <div class="clickLeft"></div>
+            <div class="clickRight"></div>
+            <img src="../img/velocidad_5.svg" alt="">
+        </div>
+    </div>
+    <div class="sidebar_separator"></div>
+    <div class="sidebar_button">
+        <div class="wrapper wrapperDual wrapperVertical" id="mirar" data-state="7" data-max-state="7">
+            <div class="clickUp"></div>
+            <div class="clickDown"></div>
+            <img src="../img/mirar_7.svg" alt="">
+        </div>
+    </div>
+    <div class="sidebar_button">
+        <div class="wrapper wrapperCapa">
+            <img class="capa capa1 active" src="../img/capa_active.svg" alt="">
+            <img class="capa capa2" src="../img/capa.svg" alt="">
+            <img class="capa capa3" src="../img/capa.svg" alt="">
+            <img class="capa capa4" src="../img/capa.svg" alt="">
+        </div>
+    </div>
+    <div class="sidebar_separator"></div>
+    <div class="sidebar_button delete_button">
+        <div class="wrapper wrapperSimple">
+            <img src="../img/tacho.svg" alt="">
+        </div>
+    </div>
+    <div class="sidebar_separator"></div>
+    <div class="sidebar_button share_button">
+        <div class="wrapper wrapperSimple">
+            <img src="../img/share.svg" alt="">
+        </div>
+    </div>
+    <div class="sidebar_button share_social_button">
+        <div class="wrapper wrapperSimple">
+            <img src="../img/social.svg" alt="">
+        </div>
+    </div>
+    <div class="sidebar_button chat_button">
+        <div class="wrapper wrapperSimple">
+            <svg width="44px" height="37px" viewBox="0 0 44 37" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+                <g id="Page-1" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
+                    <g id="V1-a" transform="translate(-245.000000, -584.000000)" fill-rule="nonzero" fill="#00FF00">
+                        <g id="multimedia" transform="translate(245.000000, 584.000000)">
+                            <path d="M27.0769231,13.4545455 L25.8076923,13.4545455 C25.3423077,13.4545455 24.9615385,13.8329545 24.9615385,14.2954545 C24.9615385,14.7579545 25.3423077,15.1363636 25.8076923,15.1363636 L27.0769231,15.1363636 C27.5423077,15.1363636 27.9230769,14.7579545 27.9230769,14.2954545 C27.9230769,13.8329545 27.5423077,13.4545455 27.0769231,13.4545455 Z" id="Shape"></path>
+                            <path d="M22,13.4545455 L15.6538462,13.4545455 C15.1884615,13.4545455 14.8076923,13.8329545 14.8076923,14.2954545 C14.8076923,14.7579545 15.1884615,15.1363636 15.6538462,15.1363636 L22,15.1363636 C22.4653846,15.1363636 22.8461538,14.7579545 22.8461538,14.2954545 C22.8461538,13.8329545 22.4653846,13.4545455 22,13.4545455 Z" id="Shape"></path>
+                            <path d="M35.5384615,8.40909091 L15.6538462,8.40909091 C15.1884615,8.40909091 14.8076923,8.7875 14.8076923,9.25 C14.8076923,9.7125 15.1884615,10.0909091 15.6538462,10.0909091 L35.5384615,10.0909091 C36.0038462,10.0909091 36.3846154,9.7125 36.3846154,9.25 C36.3846154,8.7875 36.0038462,8.40909091 35.5384615,8.40909091 Z" id="Shape"></path>
+                            <path d="M39.7692308,0 L11,0 C8.67307692,0 6.76923077,1.89204545 6.76923077,4.20454545 L6.76923077,13.4545455 L2.53846154,13.4545455 C1.14230769,13.4545455 0,14.5897727 0,15.9772727 L0,29.4318182 C0,30.8193182 1.14230769,31.9545455 2.53846154,31.9545455 L6.76923077,31.9545455 L6.76923077,36.1590909 C6.76923077,36.4954545 6.98076923,36.7897727 7.27692308,36.9159091 C7.40384615,36.9579545 7.48846154,37 7.61538462,37 C7.82692308,37 8.03846154,36.9159091 8.16538462,36.7897727 L13.4115385,32.1647727 L13.4538462,32.1227273 C13.5807692,31.9965909 13.75,31.9545455 14.0038462,31.9545455 L19.5038462,31.9545455 C20.9,31.9545455 22.0423077,30.8193182 22.0423077,29.4318182 L22.0423077,23.5454545 L22.8038462,23.5454545 C23.4384615,23.5454545 24.0307692,23.7556818 24.4961538,24.2181818 L32.4076923,31.4920455 C32.5769231,31.6181818 32.7884615,31.7022727 33,31.7022727 C33.1269231,31.7022727 33.2115385,31.6602273 33.3384615,31.6181818 C33.6346154,31.4920455 33.8461538,31.1977273 33.8461538,30.8613636 L33.8461538,23.5454545 L39.7692308,23.5454545 C42.0961538,23.5454545 44,21.6534091 44,19.3409091 L44,4.20454545 C44,1.89204545 42.0961538,0 39.7692308,0 Z M20.35,29.4318182 C20.35,29.8943182 19.9692308,30.2727273 19.5038462,30.2727273 L14.0038462,30.2727273 C13.3269231,30.2727273 12.7346154,30.4829545 12.2692308,30.9034091 L8.46153846,34.2670455 L8.46153846,31.1136364 C8.46153846,30.6511364 8.08076923,30.2727273 7.61538462,30.2727273 L2.53846154,30.2727273 C2.07307692,30.2727273 1.69230769,29.8943182 1.69230769,29.4318182 L1.69230769,15.9772727 C1.69230769,15.5147727 2.07307692,15.1363636 2.53846154,15.1363636 L6.76923077,15.1363636 L6.76923077,19.3409091 C6.76923077,21.6534091 8.67307692,23.5454545 11,23.5454545 L20.35,23.5454545 L20.35,29.4318182 Z M42.3076923,19.3409091 C42.3076923,20.7284091 41.1653846,21.8636364 39.7692308,21.8636364 L33.8461538,21.8636364 L33.8461538,19.3409091 C33.8461538,18.8784091 33.4653846,18.5 33,18.5 C32.5346154,18.5 32.1538462,18.8784091 32.1538462,19.3409091 L32.1538462,28.9272727 L25.6807692,22.9568182 C24.9192308,22.2420455 23.9038462,21.8636364 22.8038462,21.8636364 L11,21.8636364 C9.60384615,21.8636364 8.46153846,20.7284091 8.46153846,19.3409091 L8.46153846,4.20454545 C8.46153846,2.81704545 9.60384615,1.68181818 11,1.68181818 L39.7692308,1.68181818 C41.1653846,1.68181818 42.3076923,2.81704545 42.3076923,4.20454545 L42.3076923,19.3409091 Z" id="Shape"></path>
+                        </g>
+                    </g>
+                </g>
+            </svg>
+        </div>
+    </div>
+</div>
