--- public/js/index.js
@@ -1,3 +1,0 @@
-$(function() {
-    // $("#modalIndex").css("height", windowHeight);
-});
