--- sidebar.inc.html
@@ -1,76 +1,0 @@
-<!-- Sidebar -->
-<div id="sidebar">
-    <a id="btnHome" href="/">
-        <img src="../img/home.svg" />
-    </a>
-    <div class="sidebar_button_color" id="color">
-        <div class="drop_color">
-            <div class="colors-wrap">
-                <div class="color_button" data-color="#000000"></div>
-                <div class="color_button" data-color="#ffffff"></div>
-                <div class="color_button" data-color="#ff0000"></div>
-                <div class="color_button" data-color="#ff4800"></div>
-                <div class="color_button" data-color="#fff200"></div>
-                <div class="color_button" data-color="#00ff6d"></div>
-                <div class="color_button" data-color="#0049ff"></div>
-                <div class="color_button" data-color="#AE00ff"></div>
-            </div>
-        </div>
-    </div>
-    <div class="sidebar_weight_button" id="weight">
-        <div class="weight_point weight_point_tam02"></div>
-        <div class="drop_weight">
-            <div class="weight_wrap">
-                <div class="weight_button weight_button_01"></div>
-                <div class="weight_button weight_button_02"></div>
-                <div class="weight_button weight_button_03"></div>
-                <div class="weight_button weight_button_04"></div>
-            </div>
-        </div>
-    </div>
-    <div class="sidebar_button">
-        <div class="wrapper wrapperSimple" id="loop">
-            <img src="../img/girar_1.svg" alt="">
-        </div>
-    </div>
-    <div class="sidebar_button">
-        <div class="wrapper wrapperDual wrapperHorizontal" id="velocidad" data-state="5" data-max-state="10">
-            <div class="clickLeft"></div>
-            <div class="clickRight"></div>
-            <img src="../img/velocidad_5.svg" alt="">
-        </div>
-    </div>
-    <div class="sidebar_separator"></div>
-    <div class="sidebar_button">
-        <div class="wrapper wrapperDual wrapperVertical" id="mirar" data-state="7" data-max-state="7">
-            <div class="clickUp"></div>
-            <div class="clickDown"></div>
-            <img src="../img/mirar_7.svg" alt="">
-        </div>
-    </div>
-    <div class="sidebar_button">
-        <div class="wrapper wrapperCapa">
-            <img class="capa capa1 active" src="../img/capa_active.svg" alt="">
-            <img class="capa capa2" src="../img/capa.svg" alt="">
-            <img class="capa capa3" src="../img/capa.svg" alt="">
-            <img class="capa capa4" src="../img/capa.svg" alt="">
-        </div>
-    </div>
-    <div class="sidebar_separator"></div>
-    <div class="sidebar_button delete_button">
-        <div class="wrapper wrapperSimple">
-            <img src="../img/tacho.svg" alt="">
-        </div>
-    </div>
-    <div class="sidebar_separator"></div>
-    <div class="sidebar_button share_button">
-        <div class="wrapper wrapperSimple">
-            <img src="../img/share.svg" alt="">
-        </div>
-    </div>
-    <div class="sidebar_button share_social_button">
-        <div class="wrapper wrapperSimple">
-            <img src="../img/social.svg" alt="">
-        </div>
-    </div>
-</div>
